requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.2.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.2
numpy==1.26.3
